pyproject.toml
[tool.poetry]
name = "minimal-newsl"
version = "0.1.0"
description = ""
authors = ["Julio Rafael <juliorafaelnft@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "4.1.3"
python-dotenv = "1.0.1"
psycopg2-binary = "2.9.10"
pillow = "11.1.0"
crewai = {extras = ["tools"], version = "^0.102.0"}
crewai-tools = "^0.36.0"
langchain = "^0.3.18"
markdown2 = "^2.5.3"
openai = "^1.63.0"
langchain-community = "^0.3.17"
duckduckgo-search = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"